requirements.txt
snowflake-connector-python[pandas]
pandas
openpyxl
PyYAML
